backend/pyproject.toml
[project]
name = "baicore-trader-backend"
version = "1.0.0"
description = "BAI Core Trader — production trading bot backend"
authors = [{ name = "BAI Core", email = "info@baicore.kz" }]
readme = "../README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "aiosqlite>=0.20,<1.0",
    "alembic>=1.13,<2.0",
    "cryptography>=43.0,<46.0",
    "argon2-cffi>=23.1,<25.0",
    "pyjwt[crypto]>=2.9,<3.0",
    "python-multipart>=0.0.20,<1.0",
    "platformdirs>=4.3,<5.0",
    "websockets>=13.1,<15.0",
    "httpx>=0.27,<1.0",
    "python-dateutil>=2.9,<3.0",
    "apscheduler>=3.10,<4.0",
    "slowapi>=0.1.9,<1.0",
    "pandas>=2.2,<3.0",
    "numpy>=1.26,<3.0",
    "scipy>=1.13,<2.0",
    "yfinance>=0.2.50,<1.0",
    "optuna>=4.0,<5.0",
    "ccxt>=4.4,<5.0",
    "python-telegram-bot>=21.7,<22.0",
    "pywebpush>=2.0,<3.0",
    "qrcode[pil]>=8.0,<9.0",
    "pyngrok>=7.2,<8.0",
    "MetaTrader5>=5.0.45 ; sys_platform == 'win32'",
    "matplotlib>=3.9,<4.0",
    "seaborn>=0.13,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-mock>=3.14,<4.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
]

desktop = [
    "pywebview>=5.3,<6.0 ; sys_platform == 'win32'",
    "pywebview>=5.3,<6.0 ; sys_platform == 'darwin'",
    "pywebview[qt]>=5.3,<6.0 ; sys_platform == 'linux'",
    "win10toast>=0.9 ; sys_platform == 'win32'",
    "pync>=2.0 ; sys_platform == 'darwin'",
]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
